+++ .bb/skills/verify-bb/SKILL.md
@@ -123,7 +123,7 @@ and verify the processes still belong to this checkout.
 
 For worktree startup changes, use the same fresh data marker and port ownership
 checks, but run `pnpm prepare:start --worktree` followed by `pnpm start:worktree`.
-Verify the start reuses valid artifacts without invoking Turbo.
+Verify the start restores unchanged artifacts through Turbo cache hits.
 The app is served at the server URL with no Vite listener. Record preparation
 and time to server/daemon health independently; repeat preparation to verify
 Turbo hits. Check code, asset, skill, manifest, SDK/toolchain, and staged-asset
@@ -135,7 +135,7 @@ Verify missing/stale preparation is rebuilt automatically before services start,
 and native modules are checked and repaired when necessary.
 Preparation writes build outputs, so use a separate staging checkout to warm
 cache while a live instance still serves those files. Keep the serving checkout
-path stable. See `docs/debugging-and-qa.md` for the preparation receipt contract.
+path stable. See `docs/debugging-and-qa.md` for the preparation and cleanup contract.
 Stop only verified PIDs spawned by this test; `bb-dev-app stop` is for instances
 managed by that launcher, not arbitrary `pnpm start:worktree` processes.
